fiubar/requirements.txt
django-contrib-comments
django-markup-deprecated
django-registration-redux
Pillow
python-dateutil